candidate for live
